requirements.txt
# Prime — Production AI Agent Framework
# Core (required)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiohttp>=3.9.0

# Database
aiosqlite>=0.20.0

# Telegram
# python-telegram-bot>=21.0  # optional, we use direct API calls

# Web search (optional, falls back to curl)
duckduckgo-search>=6.0.0

# Utilities
pydantic>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Scheduling (optional, for background tasks)
apscheduler>=3.10.0

# Dev/testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
